backend_public/pyproject.toml
[tool.poetry]
name = "auditor-backend-public"
version = "0.1.0"
description = "Public FastAPI backend for AI prompt auditor metrics and workflow reporting"
authors = ["Your Team <dev@example.com>"]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
python-dotenv = "^1.0.1"
SQLAlchemy = "^2.0.31"
asyncpg = "^0.29.0"
pydantic = "^2.7.4"
pydantic-settings = "^2.3.3"
python-jose = {extras=["cryptography"], version="^3.3.0"}
passlib = {extras=["bcrypt"], version="^1.7.4"}
authlib = "^1.3.1"
httpx = "^0.27.0"
python-multipart = "^0.0.9"
itsdangerous = "^2.2.0"
sqlmodel = "^0.0.21"
alembic = "^1.13.1"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
ruff = "^0.4.8"
pytest = "^8.2.0"
pytest-asyncio = "^0.23.8"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
